\name{bar_chart}
\alias{bar_chart}
\title{generate an interactive bar chart}
\usage{
  bar_chart(df, target_dir = "bar_chart",
    path_to_template = "templates/bar-chart.zip")
}
\arguments{
  \item{base}{a base note for the chord}
}
\value{
  a folder with an interactive bar chart in it
}
\description{
  generate an interactive bar chart
}
\examples{
#not run
}

