\name{charrrt_setup}
\alias{charrrt_setup}
\title{setup the output folder/file structure for any given chart}
\usage{
  charrrt_setup(output_directory)
}
\arguments{
  \item{output_directory}{the name of the folder to}
}
\value{
  a folder ready for an interactive chart and data file to
  be plopped in
}
\description{
  setup the output folder/file structure for any given
  chart
}
\examples{
#not run
}

