\name{highChaRRRts}
\alias{highChaRRRts}
\title{generate a set of interactive high Charts}
\usage{
  highChaRRRts(dcasted_df, type = "bearclaw", main = "",
    xlab = "", ylab = "", output = "highChaRRRt",
    pal = NULL)
}
\arguments{
  \item{df}{your dcasted dataframe}

  \item{type}{either line, bar, area, column}

  \item{main}{the title of the chart}

  \item{xlab}{the chart's x label}

  \item{ylab}{the chart's y label}
}
\value{
  a folder with the name determined from output with an
  interactive bar chart of your data in it
}
\description{
  generate a set of interactive high Charts
}
\examples{
#not run
}

