\name{highChaRRRts_setup}
\alias{highChaRRRts_setup}
\title{setup the output folder/file structure for any given chart}
\usage{
  highChaRRRts_setup(chart_type)
}
\arguments{
  \item{base}{a base note for the chord}
}
\value{
  a folder ready for an interactive chart and data file to
  be plopped in
}
\description{
  setup the output folder/file structure for any given
  chart
}
\examples{
#not run
}

